+++ lib/shape/shape.dart
@@ -1,0 +1,14 @@
+export 'base.dart';
+export 'capsule.dart';
+export 'circle.dart';
+export 'cube.dart';
+export 'cuboid.dart';
+export 'cylinder.dart';
+export 'ellipse.dart';
+export 'ellipsoid.dart';
+export 'rectangle.dart';
+export 'sphere.dart';
+export 'square.dart';
+export 'stadium.dart';
+export 'triangle.dart';
+export 'triangle_right.dart';